Moves 'execute' package to a module in kcli dir
