--- tools/kcli/kcli/execute/__init__.py
@@ -1,2 +1,0 @@
-# from kcli.execute import core
-# from kcli.execute import execute
